Creating word vectors with word2vec
I ran the cells which were commented out and then re-commented them, e.g. training the actual model and then saving it. I also re-loaded the model and it works fine.

Only real changes were the addition of the keyed vector `.wv` to a lot of the lookup functions. This separates the read-only keyed vector lookup operations from training. It was throwing a lot of deprecation errors.